feat: schema file hot reload
